Fix deblock_qed
